Take another callback to give progress on the file itself.
2002-06-12  Joe Shaw  <joe@ximian.com>

	* rcd-fetch.c (rcd_fetch_packages): Take another callback to give
	progress on the file itself.

	* rcd-pending.c (rcd_pending_get_user_data,
	rcd_pending_set_user_data): Remove these... it's a GObject, so we
	can just use g_object_{set|get}_data().
	(rcd_pending_add_message, rcd_pending_get_messages): Added.
	Allows persistent messages to be set on an RCDPending so important
	info isn't lost in between polls.

	* rcd-rpc-packsys.c: Add messages to the pending object when we do
	important things, update it in package download progress callback.

	* rcd-rpc-system.c (system_poll_pending): Serialize the messages.

	* rcd-transfer.c: Use gsize for file sizes throughout.

+++ src/rcd-transfer.c
@@ -270,7 +270,7 @@ rcd_transfer_set_error(RCDTransfer *t, RCDTransferError err,
 } /* rcd_transfer_set_error */
 
 static void
-rcd_transfer_emit_data(RCDTransfer *t, char *buf, int size)
+rcd_transfer_emit_data(RCDTransfer *t, char *buf, guint64 size)
 {
     t->trans_size += size;
 
@@ -455,7 +455,7 @@ file_read_data(GIOChannel *iochannel,
     RCDTransfer *transfer = data;
     GIOError err;
     char buf[BLOCK_SIZE];
-    int bytes;
+    gsize bytes;
     int watch;
 
     watch = GPOINTER_TO_INT(transfer->proto_data);
